Fix some deadlocks in the case of a comm failure when smart_poll() is called without holding the UPS lock. Contributed by Keith Campbell <campbell@econnectix.com>:
On one of our machines, apcupsd 3.14.5 (using a smart serial cable)
would occasionally not finish starting up.  Running strace revealed
that it was stuck in deadlock on a futex() system call.

I've traced this to some poorly written code inside the apcsmart
driver.  Inside smart.c there is a UPSlinkCheck() function which
according to the documentation must be called with the UPS lock held
(ugh).  smart_poll() calls the UPSlinkCheck() function (but only in a
rarely followed error handling branch), so therefore it must also be
called with the lock held.  The problem is, many things call
smart_poll() but some call it without the lock, and other with the
lock.  I've put locks around a few smart_poll() calls in smartsetup2.c
which corrected the problem I was seeing.

--HG--
branch : Branch-3_14